refactor: rename setup to config command
