+++ internal/cmd/config/setup.go
@@ -1,4 +1,4 @@
-package setup
+package config
 
 import (
 	"fmt"
